Start writing stack implementation
I thought about this a little and I'm pretty sure it's going to be
better to go through the pain of implementing push, insert, peek, pop
and remove operations separately for each primitive type, because it
will make the library so much more comfortable to use in the future.

I might change my mind mid-implementation though, we'll see.