[stack] Finish all function implementations
I'll definitely play around with them still, and they have yet to be
tested, so this version is not final.

I slightly altered the suffix naming convention, now functions that
operate on void* have suffix "p", as opposed to no suffix at all, e.g.:

	rnd_stack_push() -> rnd_stack_pushp()

This makes the intent more clear, and the naming more consistent.