Pointer functions revamp + other fixes
- renamed RND_ILLEGAL to proper RND_EILLEGAL (original was a mistake)
- renamed all *p functions back to the version without p suffix (this
  reverts part of the changes made by
  17d77bd77e2429b25be2156d6579fe643b273af1)
- fix functions that operate directly on void*
- add insert/remove testing unit (passing)

The former code for pointer functions was a bit of a mess, because, even
though I didn't realize it at the time, I didn't know exactly what I
wanted the pointer functions to do (operate on arbitrary data, or
operate on void*, which hide arbitrary data?).

The current behavior is: Each function that operates directly on void*
expects arbitrary data or size elem_size, which starts at the void* and
spans adjacent bytes (if any).
Rationale:
	- it's better to lay things in contiguous blocks, rather than
	  jump in memory
	- you can still store void*, if you really want to (just make a
	  stack with elem_size = sizeof(void*) and then pass void** to
	  rnd_stack_push(), etc.)