[stack] Add quick insert and remove operations
quickinsert and quickremove achieve the same goal as insert and remove,
but they work in O(1) by swapping in/out a single element, rather than
shifting entire memory blocks. This means they change the order of
elements, and should only be used if it's not important to preserve the
order.

More precisely, quickinsert works by swapping the element residing at
our "target spot" to the top of the stack. Here's a comparison of
regular and quick insert:

regular insert:
	1 2 3 4 _ _ _ (top: 4, bottom: 1)
	<insert 5 at position 2>
	1 2 5 3 4 _ _ (top: 4, bottom: 1)
	    ^
quick insert:
	1 2 3 4 _ _ _ (top: 4, bottom: 1)
	<quickinsert 5 at position 1>
	1 2 5 4 3 _ _ (top: 3, bottom: 1)
	    ^

Quickremove works in the same way, by swapping the top of the stack into
the spot we want to remove. Here's a comparison of regular and quick
remove:

regular remove:
	1 2 3 4 5 _ _ _ (top: 5, bottom: 1)
	<remove position 2>
	1 2 4 5 _ _ _ _ (top: 5, bottom: 1)

quick remove:
	1 2 3 4 5 _ _ _ (top: 5, bottom: 1)
	<quickremove position 2>
	1 2 5 4 _ _ _ _ (top: 4, bottom: 1)