[stack] Change some warnings into errors
In debug mode, it's no problem if you peek or pop when size is 0, but
outside of debug mode, it's undefined behavior. Debug mode should give
feedback as though the code is running without debug mode, so
peeking/popping with size 0 should throw errors, not warnings.

An alternative would be always checking for size==0 case, even outside
debug mode, and I'm still considering it so it might change in the
future. I'm unsure whether popping/peeking with size 0 is something
that's convenient to be able to do in programs, or an indication of a
bug.