Update planned implementations lists
I got rid of "array", because it's provided by the C language itself,
there's no need for another wrapper type. And it would be a bit
ambiguous to implement, e.g. how would it differ from the stack
implementation?

Added deque, because having stack and queue separately begs for a deque
type.