Copy queue files from existing stack files
Queue and stack are very similar, and the implementation will reflect
that as well, which is why I won't be writing queue from scratch; rather
I'll take existing stack code and morph it into queue code. The first
step is taking a raw copy, which is what this commit does (I substituted
"stack" to "queue" in both files though).