[stack] Remove +64 from stack_copy
Originally, when I wrote rnd_stack_copy, I thought that it would be a
good idea to increase the size of the destination stack when copying, to
leave some leeway for a handful of pushes/pops without need for another
realloc.

Now I think that idea was stupid, because it runs into the complication
of "what if +64 causes a size_t overflow?". I don't want to make the
function unnecessarily complex or clever - it should be dumb and simple,
so that the user can leverage it more effectively.