[queue] Add push functions implementations
Man page added also.

A new function has been added to helpers.c, rnd_ringbuffit for the same
purpose as rnd_buffit, but working on ring buffers. In ring buffers, it
is necessary to perform some additional cleaning up after a buffer
resize, and the pointers to head and tail need to be updated after
realloc.