[queue] Simplify example in quickinsert man page
At first I wanted to make the example more interesting by demonstrating
how the ring buffer looks like in memory, but on second thought, it only
pollutes the example with unnecessary complications. As far as the user
is concerned, it doesn't matter how the queue is laid in memory, unless
they want to extend it themselves (but then they can read rnd_queue(3)).