Replace suffixed man pages with NAME section lists
(follow-up to a19ba134de3b671fcd97430)

I didn't realize it at the time, but man pages have a built-in mechanism
that can be used for aliasing: all the names that are listed in the NAME
section of a man page will be used as keywords for matching. This is
great, because it allows me to cut down the number of man pages
drastically.

IMPORTANT NOTE:
If you have installed versions of the library from commit
a19ba134de3b671fcd97430 onwards, run `make uninstall` BEFORE checking
out this commit. Otherwise you will have a lot of redundant man pages
left in /usr/local/man/man3.