[queue] Move dtor example from destroy to clear
Now both stack and queue show the dtor example in their respective clear
function man pages, and the destroy man pages redirect to the clear
ones.

Also, the queue dtor example used "s" as the rnd_queue variable name, so
I changed it to "q" to match. Cosmetic change, but it matters for
readability.