Improve test automation in Makefile
Now there is a single "test_%" target for testing all modules, and
there will be no need to implement a new test-stack, test-queue, etc.
targets whenever a new module is created. There is a new MODULES macro
that needs to contain the names of all modules for that to work though.

Added .SECONDARY to disable the removal of intermediary files by make.
This improves compilation speeds of the test target.

Added test_clean target for cleaning test/obj and test/bin. This makes
it possible to run "make clean" without deleting test build files, which
take a long time to recompile.

Minor syntax improvements (added -- arguments to some commands).