Bump to 2.0.2
The only change is the license, but I kind of have to draw a firm line
by changing the version number, because otherwise things could get
pretty vague in the future.

Release 2.0.1 is functionality-wise identical to 2.0.2, but bears a
different license (GPLv2). This makes the functionality from 2.0.1
available under 2 different licenses, but that's fine, especially since
LGPL is more permissive than GPL.