Revert adding getptr function
This reverts commit b95384bd719e2c072240d15f70e9585badb44b4c.

I'm backpedaling on the getptr function, because I realized it's
probably better to change the existing functionality of generic get and
peek functions to do what getptr was doing.

Still, I don't want to lose the getptr implementation just in case, so
I'm keeping it in history for potential future reference.