Fix redundant file outputs in generate.lua
I'm in too much hurry to explain, but basically the old script would
output the same files many times instead of doing each file one time.

It's now been fixed. One of the things that needed to be changed is
adding an extra parameter to C_ERR_INCOMPAT_ELEM_TYPE.