Remove pconf sorting for a simpler solution
The sorting had an issue: the order wasn't the same as the one in the
generate script, which made for annoying results. And it was only needed
for filtering psets inside FDef.lua, so I removed the functionality
altogether and opted for a simpler, one-shot solution inside FDef.lua.

Now, the order is kept the same as the one chosen by the user inside
ParamConfig:new().