Add automatic testing via GitHub Actions
A push to any branch will trigger the build and execution of a
Dockerfile which tests the library when built with GCC or clang
(may extend to more compilers in the future).
Unfortunately, for whatever reason valgrind refused to work inside
docker, so I disabled it (inside docker only; it will still be used when
running `make test` by hand). Perhaps the alpine base image is at fault
here, I may experiment with something more mainstream like debian later.