feat(utils): Add boolean macros
For SP_SIZEOF_BOOL I could ask the user to just write 0, but this is
more readable in-code, and makes it possible to alter the API without
breaking user code in the future (although I highly, *highly* doubt it
will happen).

SP_BYTE_SIZE has similar rationale, I'm not yet sure if keeping the code
generalized for an arbitrary CHAR_BIT will be feasible, so if worst
comes to worst, I can just change it to 8 in a single place. And if it
will be feasible, it's nice to expose SP_BYTE_SIZE in case the user
wants to know for some secret sauce optimization.

Issue: #6