refactor: Add DATA_SIZE macro
The DATA_SIZE macro aims to simplify away a lot of the new logic that
might be needed to deal with boolean data structures (due to elem_size
being 0, which means the old size * elem_size formula doesn't work
all the time anymore).
Also, it improves readability by quite a bit, which is a nice bonus.

I believe this also fixes a bug in sp_stack_copy and sp_queue_copy when
copying between structures with varying elem_sizes - I would normally
write a test to prevent this but it's not easy to detect, since it deals
with malloc'd buffer sizes which are not explicitly stored anywhere
(instead are inferred from size and elem_size).

Issue: #6